Consolidate full screen and modal flows
